Conflicts:
	UPDATE.sql
	www/modules/module_xpay/module_xpay.class.php
	www/modules/module_xpay/templates/actions/do_payment.tpl
	www/modules/module_xpay_cielo/includes/module_xpay_cielo.pedido.model.php
	www/modules/module_xpay_cielo/module_xpay_cielo.class.php

+++ www/modules/module_xclasse/module_xclasse.class.php
@@ -1,38 +1,43 @@
 <?php
 /*
-class MagesterCourseTree extends MagesterTree {
+class MagesterCourseTree extends MagesterTree
+{
 	public function insertNode($node, $parentNode = false, $previousNode = false) {}
 	public function removeNode($node) {}
 	public function reset() {}
 }
 */
 
-class module_xclasse extends MagesterExtendedModule {
-	
+class module_xclasse extends MagesterExtendedModule
+{
     // Mandatory functions required for module function
-    public function getName() {
+    public function getName()
+    {
         return "XCLASSE";
     }
-    
-    public function getPermittedRoles() {
+
+    public function getPermittedRoles()
+    {
         return array("administrator");
     }
 
-    public function isLessonModule() {
+    public function isLessonModule()
+    {
         return false;
     }
-    
-    public function getUrl($action) {
-    	switch($action) {
+
+    public function getUrl($action)
+    {
+    	switch ($action) {
     		case "new_course_class" : {
-    			return $this -> moduleBaseUrl . 
-    				"&action=" . $action . 
+    			return $this -> moduleBaseUrl .
+    				"&action=" . $action .
     				"&xcourse_id=" . $this->getEditedCourse()->course['id'];
     		}
 	   		case "edit_course_class" : {
-    			return $this -> moduleBaseUrl . 
-    				"&action=" . $action . 
-    				"&xcourse_id=" . $this->getEditedCourse()->course['id'] . 
+    			return $this -> moduleBaseUrl .
+    				"&action=" . $action .
+    				"&xcourse_id=" . $this->getEditedCourse()->course['id'] .
     				"&xclasse_id=" . $this->getEditedCourseClass()->classe['id'];
     		}
     		default : {
@@ -40,8 +45,9 @@ class module_xclasse extends MagesterExtendedModule {
     		}
     	}
     }
-    public function getTitle($action) {
-    	switch($action) {
+    public function getTitle($action)
+    {
+    	switch ($action) {
     		case "new_course_class" : {
     			return __XCLASSE_NEWCLASS;
     		}
@@ -57,55 +63,19 @@ class module_xclasse extends MagesterExtendedModule {
     	}
     }
 
-    public function getDefaultAction() {
+    public function getDefaultAction()
+    {
     	return "view_course_class";
     }
-	
+
     /* ACTIONS FUNCTIONS */
-    public function newCourseClassAction() {
+    public function newCourseClassAction()
+    {
     	$smarty 		= $this->getSmartyVar();
     	$currentUser	= $this->getCurrentUser();
-    	
+
     	$xuserModule = $this->loadModule("xuser");
-    	
-		if ($xuserModule->getExtendedTypeID($currentUser) != 'administrator' &&
-			$xuserModule->getExtendedTypeID($currentUser) != 'coordenator'
-		) {
-			header("Location: " . $this->moduleBaseUrl);
-			exit;
-		}
-		
-		if ( $this->makeClassesForm() ) {
-			return true;
-       	} else {
-       		return false;
-       	}
-    }
-    public function editCourseClassAction() {
-    	$smarty 		= $this->getSmartyVar();
-    	$currentUser	= $this->getCurrentUser();
-    	
-    	$xuserModule = $this->loadModule("xuser");
-    	
-		if ($xuserModule->getExtendedTypeID($currentUser) != 'administrator' &&
-			$xuserModule->getExtendedTypeID($currentUser) != 'coordenator'
-		) {
-			header("Location: " . $this->moduleBaseUrl);
-			exit;
-		}
-		
-		if ( $this->makeClassesForm() ) {
-			return true;
-       	} else {
-       		return false;
-       	}
-    }
-    public function deleteCourseClassAction() {
-    	$smarty 		= $this->getSmartyVar();
-    	$currentUser	= $this->getCurrentUser();
-    	
-    	$xuserModule = $this->loadModule("xuser");
-    	
+
 		if ($xuserModule->getExtendedTypeID($currentUser) != 'administrator' &&
 			$xuserModule->getExtendedTypeID($currentUser) != 'coordenator'
 		) {
@@ -113,58 +83,100 @@ class module_xclasse extends MagesterExtendedModule {
 			exit;
 		}
 
-		// 
+		if ( $this->makeClassesForm() ) {
+			return true;
+       	} else {
+       		return false;
+       	}
+    }
+    public function editCourseClassAction()
+    {
+    	$smarty 		= $this->getSmartyVar();
+    	$currentUser	= $this->getCurrentUser();
+
+    	$xuserModule = $this->loadModule("xuser");
+
+		if ($xuserModule->getExtendedTypeID($currentUser) != 'administrator' &&
+			$xuserModule->getExtendedTypeID($currentUser) != 'coordenator'
+		) {
+			header("Location: " . $this->moduleBaseUrl);
+			exit;
+		}
+
+		if ( $this->makeClassesForm() ) {
+			return true;
+       	} else {
+       		return false;
+       	}
+    }
+    public function deleteCourseClassAction()
+    {
+    	$smarty 		= $this->getSmartyVar();
+    	$currentUser	= $this->getCurrentUser();
+
+    	$xuserModule = $this->loadModule("xuser");
+
+		if ($xuserModule->getExtendedTypeID($currentUser) != 'administrator' &&
+			$xuserModule->getExtendedTypeID($currentUser) != 'coordenator'
+		) {
+			header("Location: " . $this->moduleBaseUrl);
+			exit;
+		}
+
+		//
 		if ($this->getEditedCourseClass()) {
-			
+
 			$this->getEditedCourseClass()->delete();
-			
+
 			$redirect = "administrator.php?ctg=courses&course=" . $this->getEditedCourse() -> course['id'] . "&op=course_classes&message=".urlencode(__XCLASSE_DELETE_MESSAGE)."&message_type=success";
 		} else {
 			$redirect = "administrator.php?ctg=courses&course=" . $this->getEditedCourse() -> course['id'] . "&op=course_classes&message=".urlencode(__XCLASSE_ERROR_DELETE_MESSAGE)."&message_type=failure";
 		}
-		
+
 		eF_redirect($redirect);
 		exit;
-    }    
-    
-    public function getClassSchedulesAction($token = null, $fields = null) {
+    }
+
+    public function getClassSchedulesAction($token = null, $fields = null)
+    {
         $smarty = $this -> getSmartyVar();
 		$smarty -> assign("T_MODULE_XCOURSE_ACTION", $selectedAction);
-		
-		// LOAD L10N DATA	
+
+		// LOAD L10N DATA
         $modules = eF_loadAllModules(true);
         $l10nSection = $modules['module_language']->getSection("l10n");
-        	
+
         $smarty -> assign("T_L10N_DATA", $l10nSection['data']);
-        	
+
         // LOAD SCHEDULES FOR CLASS
 		$courseClasses = $this->getEditedCourse()->getCourseClasses();
-			
+
 		if (eF_checkParameter($_GET['xcourse_class_id'], 'id')) {
 			$courseClassID = $_GET['xcourse_class_id'];
-				
+
 			$smarty -> assign("T_XCOURSE_CLASS_SCHEDULES", $courseClasses[$courseClassID]->classe['schedules']);
-				
+
 			$smarty -> assign("T_FORM_ACTION", $this->moduleBaseUrl . '&action=' . self::SAVE_CLASS_SCHEDULES . '&xcourse_class_id=' . $courseClassID);
-				
+
 			echo $result = $smarty->fetch(
 				$this->moduleBaseDir . '/templates/actions/' . self::GET_CLASS_SCHEDULES . '.tpl'
 			);
 		}
 		exit;
     }
-    public function saveClassSchedulesAction() {
+    public function saveClassSchedulesAction()
+    {
 		if (eF_checkParameter($_GET['xcourse_class_id'], 'id')) {
 			$insertData = array(
 				'week_day' 	=> $_POST['week_day']['new'],
 				'start'		=> $_POST['start']['new'],
 				'end' 		=> $_POST['end']['new']
 			);
-			
+
 			unset($_POST['week_day']['new']);
 			unset($_POST['start']['new']);
 			unset($_POST['end']['new']);
-			
+
 			$updateData = array(
 				'week_day' 	=> $_POST['week_day'],
 				'start'		=> $_POST['start'],
@@ -173,26 +185,25 @@ class module_xclasse extends MagesterExtendedModule {
 
 			$courseClassID = $_GET['xcourse_class_id'];
 			$courseClass = new MagesterCourseClass($courseClassID);
-			
+
 			$courseClass->clearSchedule();
-			
-			
-			foreach($updateData['week_day'] as $index => $value) {
+
+			foreach ($updateData['week_day'] as $index => $value) {
 				$courseClass->appendSchedule(
 					$updateData['week_day'][$index],
 					$updateData['start'][$index],
 					$updateData['end'][$index]
 				);
 			}
-			
-			foreach($insertData['week_day'] as $index => $value) {
+
+			foreach ($insertData['week_day'] as $index => $value) {
 				$courseClass->appendSchedule(
 					$insertData['week_day'][$index],
 					$insertData['start'][$index],
 					$insertData['end'][$index]
 				);
 			}
-			
+
 			if ($courseClass->persistSchedule()) {
 				echo json_encode(array(
 					'message'		=> __XCOURSE_CLASS_SCHEDULE_SAVE_MESSAGE,
@@ -212,9 +223,9 @@ class module_xclasse extends MagesterExtendedModule {
 		}
 		exit;
     }
-	
 
-	public function editXcourseAction() {
+	public function editXcourseAction()
+	{
        	if ( $this->makeCourseClassesForm() ) {
 			$this->appendTemplate(array(
            		'title'			=> __XCOURSE_EDITXCOURSECLASSES,
@@ -225,40 +236,33 @@ class module_xclasse extends MagesterExtendedModule {
        	}
        	$this->addModuleData('edited_course', $this->getEditedCourse()->course);
 	}
-	
-	
-	
-	
-	
 
-    
     /* MAIN-INDEPENDENT MODULE INFO, PAGES, TEMPLATES, ETC... */
     /* UTILITY FUNCTIONS */
 
-    private function makeClassesForm() {
-    	
+    private function makeClassesForm()
+    {
     	$smarty = $this->getSmartyVar();
-    	
+
     	if (isset($currentUser -> coreAccess['lessons']) && $currentUser -> coreAccess['lessons'] == 'hidden') {
 			eF_redirect(basename($_SERVER['PHP_SELF'])."?ctg=control_panel&message=".urlencode(_UNAUTHORIZEDACCESS)."&message_type=failure");
-		} else if (isset($currentUser -> coreAccess['lessons']) && $currentUser -> coreAccess['lessons'] != 'change') {
+		} elseif (isset($currentUser -> coreAccess['lessons']) && $currentUser -> coreAccess['lessons'] != 'change') {
 			$_change_ = false;
 		} else {
 			$_change_ = true;
 		}
-    	
-    	
+
 		$formClass = new HTML_QuickForm(
-			"add_courseclass_form", 
-			"post", 
-			$_SERVER['REQUEST_URI'], 
+			"add_courseclass_form",
+			"post",
+			$_SERVER['REQUEST_URI'],
 			"", null, true
 		);
-		
+
 		$formClass -> addElement('hidden', 'coursesID');
 		$formClass -> addElement('hidden', 'action');
 		$formClass -> addElement('hidden', 'id');
-		
+
 	 	$formClass -> addElement('text', 'name', _COURSECLASSNAME, 'class = "large"');
 	 	$formClass -> addRule('name', _THEFIELD.' "'._COURSECLASSNAME.'" '._ISMANDATORY, 'required', null, 'client');
 	 	$formClass -> addElement('advcheckbox', 'active', _ACTIVEFEM, null, null, array(0, 1));
@@ -266,18 +270,18 @@ class module_xclasse extends MagesterExtendedModule {
 
 	 	$formClass -> addElement('text', 'max_users', _MAXSTUDENTS, 'style = "display: none;"');
 	 	$formClass -> addRule('max_users', _THEFIELD.' "'._MAXSTUDENTS.'" '._ISMANDATORY, 'numeric', null, 'client');
-	 	
+
 	 	$formClass -> addElement('jquerydate', 'start_date', _STARTDATE);
 	 	$formClass -> addElement('jquerydate', 'end_date', _ENDDATE);
-	 	
+
 		$formClass -> addElement('submit', 'submit_xclasse_form', _SUBMIT);
-		
+
 		if (!$_change_) {
 			$formClass -> freeze();
 		} else {
 			if ($formClass -> isSubmitted() && $formClass -> validate()) {
 				$fields = $formClass -> exportValues();
-						
+
 				$fields = array(
 					'courses_ID'		=> $formClass -> exportValue('coursesID') != 0 ? $formClass -> exportValue('coursesID') : $this->getEditedCourse()->course['id'],
 					'id'				=> $formClass -> exportValue('id'),
@@ -286,7 +290,7 @@ class module_xclasse extends MagesterExtendedModule {
 					'active'			=> $formClass -> exportValue('active'),
 					'duration'			=> 0,
 					'options'			=> '',
-					'languages_NAME'	=> $GLOBALS['configuration']['default_language'], 
+					'languages_NAME'	=> $GLOBALS['configuration']['default_language'],
 					'metadata'			=> null,
 					'share_folder'		=> null,
 					//'created'			=> time(),
@@ -299,32 +303,32 @@ class module_xclasse extends MagesterExtendedModule {
 				if (date_create_from_format('d/m/Y', $formClass -> exportValue('end_date')) !== FALSE) {
 					$fields['end_date']	= date_create_from_format('d/m/Y', $formClass -> exportValue('end_date'))->format('U');
 				}
-				
+
 				if (is_numeric($fields['id']) && $fields['id'] > 0) {
-					// UPDATE CourseClass	
+					// UPDATE CourseClass
 					$editCourseClass = new MagesterCourseClass($fields['id']);
 					$editCourseClass->classe = array_merge($editCourseClass->classe, $fields);
 					$editCourseClass->persist();
 					//$redirect = $this->moduleBaseUrl . "&action=edit_course_class&xcourse_id=".$this->getEditedCourse() -> course['id']."&xclasse_id=" . $editCourseClass->classe['id'] . &message=".urlencode(__XCLASSE_UPDATE_MESSAGE)."&message_type=success";
-					
+
 					$redirect = "administrator.php?ctg=courses&course=" . $this->getEditedCourse() -> course['id'] . "&op=course_classes&message=".urlencode(__XCLASSE_UPDATE_MESSAGE)."&message_type=success";
-					
+
 				} else {
 					$editCourseClass = MagesterCourseClass:: createCourseClass($fields);
-					
+
 					//$redirect = $this->moduleBaseUrl . "&action=edit_course_class&xcourse_id=".$this->getEditedCourse() -> course['id']."&xclasse_id=" . $editCourseClass->classe['id'] . "&message=".urlencode(__XCLASSE_INSERT_MESSAGE)."&message_type=success";
-					
+
 					$redirect = "administrator.php?ctg=courses&course=" . $this->getEditedCourse() -> course['id'] . "&op=course_classes&message=".urlencode(__XCLASSE_UPDATE_MESSAGE)."&message_type=success";
 				}
-				
+
 				!isset($redirect) OR eF_redirect($redirect);
-				
+
 				//$smarty -> assign("T_REDIRECT_PARENT_TO", basename($_SERVER['PHP_SELF'])."?ctg=courses&edit_course=" . $course->course['id'] );
 				// RELOAD CLASS LIST
 			}
 		}
-		
-		
+
+
 		/*
 		$classes = $this->getEditedCourse() -> getCourseClasses($constraints);
 		$totalEntries = $this->getEditedCourse() -> countCourseClasses($constraints);
@@ -332,7 +336,7 @@ class module_xclasse extends MagesterExtendedModule {
 		*/
 		if ($this->getEditedCourseClass(true)) {
 			$courseClass = $this->getEditedCourseClass();
-			
+
 			$defaults = $courseClass->classe;
 		} else {
 			$defaults = MagesterCourseClass :: getDefaultCourseClassValues();
@@ -340,26 +344,20 @@ class module_xclasse extends MagesterExtendedModule {
 		/* GET CONFIG DATE OPTIONS */
 		$defaults['start_date']	= date('d/m/Y', $defaults['start_date']);
 		$defaults['end_date']	= date('d/m/Y', $defaults['end_date']);
-		
+
 		unset($defaults['courses_ID']);
-		
+
 		$formClass->setDefaults($defaults);
 		$formClass->setDefaults(array(
 			'coursesID'	=> $this->getEditedCourse()->course['id']
 		));
-		
+
 		$rendererClass = new HTML_QuickForm_Renderer_ArraySmarty($smarty);
 		$formClass -> accept($rendererClass);
 		$smarty -> assign('T_XCLASSE_FORM', $rendererClass -> toArray());
-		
-		$smarty -> assign ("T_XCOURSE_CLASSES_LIST", $xcourseClasses);				
-		
+
+		$smarty -> assign ("T_XCOURSE_CLASSES_LIST", $xcourseClasses);
+
 		return true;
     }
-
-   
-    
-    /* Data Model Functions */
- 
 }
-?>
